일정 관리 앱  v2.2 -> 오류 수정(updateSchedule의 반환 update("SQL문") 뒤에 ,id 추가) 및 각 클래스와 인터페이스에 주석으로 설명 추가

+++ src/main/java/com/example/calendar/service/ScheduleServiceImpl.java
@@ -20,21 +20,24 @@ import java.util.Optional;
  */
 @Service
 public class ScheduleServiceImpl implements ScheduleService { // ScheduleService 인터페이스를 구현한 구현체
+    // 데이터 접근을 위한 ScheduleRepository 인스턴스
     private final ScheduleRepository scheduleRepository;
 
+    // 생성자 주입을 통해 ScheduleRespository 주입
     public ScheduleServiceImpl(ScheduleRepository scheduleRepository) {
         this.scheduleRepository = scheduleRepository;
     }
 
     @Override
     public ScheduleResponseDto createSchedule(ScheduleRequestDto dto) {
-        // DB에 저장
+        // Schedule 데이터를 DB에 저장하고 결과 반환
         ScheduleResponseDto result = scheduleRepository.createSchedule(dto);
         return result;
     }
 
     @Override
     public List<ScheduleResponseDto> findAllSchedules(String userName, String updatedAt) {
+        // 조건에 맞는 일정 목록을 조회
         List<ScheduleResponseDto> result = scheduleRepository.findAllSchedules(userName, updatedAt);
         return result;
     }
@@ -45,7 +48,7 @@ public class ScheduleServiceImpl implements ScheduleService { // ScheduleService
         // List의 경우 데이터가 없으면 빈배열을 응답했지만 ScheduleResponseDto 객체 하나 조회의 경우
         Optional<ScheduleResponseDto> optionalSchedule = scheduleRepository.findScheduleById(id);
 
-        // NPE(NullPointerException) 방지
+        // 조회된 일정이 없을 경우 예외 발생으로 NPE(NullPointerException) 방지
         if (optionalSchedule.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 id입니다. = " + id);
         }
@@ -54,31 +57,48 @@ public class ScheduleServiceImpl implements ScheduleService { // ScheduleService
 
     @Override
     public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto dto) {
-        String password = scheduleRepository.findPasswordById(id);
+        // id로 일정 비밀번호 조회
+        Optional<String> password = scheduleRepository.findPasswordById(id);
 
         if (password.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 id입니다. = " + id);
         }
 
+        // 필수 값 누락 확인
         if (dto.getUserName() == null || dto.getPassword() == null || dto.getTitle() == null || dto.getContents() == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "필수 정보를 모두 입력하세요.");
         }
 
-        if (!dto.getPassword().equals(password)) {
+        // 비밀번호 일치 확인
+        if (!dto.getPassword().equals(password.get())) {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "올바른 비밀번호를 입력하세요."); // 401번...
         }
 
+        // 업데이트 실행 결과 확인
         int updatedRow = scheduleRepository.updateSchedule(id, dto);
 
         if (updatedRow == 0) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "수정된 데이터가 없습니다.");
         }
 
+        // 조회된 데이터를 Optional에서 꺼내어 반환
         return scheduleRepository.findScheduleById(id).get(); //  .Get() Optional에 감싸진 데이터를 꼭 꺼내오자.
     }
 
     @Override
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, String password) {
+        // 주어진 id로 일정 비밀번호 조회
+        Optional<String> dbPassword = scheduleRepository.findPasswordById(id);
+        if (dbPassword.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 id입니다. = " + id);
+        }
+
+        // 비밀번호 일치 여부 확인
+        if (!dbPassword.get().equals(password)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "올바른 비밀번호를 입력하세요."); // 401번...
+        }
+
+        // 삭제 작업 결과 확인
         int deletedRow = scheduleRepository.deleteSchedule(id);
 
         if (deletedRow == 0) {
